LayoutTests/editing/selection/shift-click-live-range-expected.txt

Tests that shift+clicking does the platform correct behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
FAIL window.getSelection().toString() is "two three
four five " and should be "two three
four five"
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
PASS window.getSelection().toString() is correct
PASS Selection direction is correct.
FAIL window.getSelection().toString() is "one two" and should be "one two "
PASS Selection direction is correct.
FAIL window.getSelection().toString() is "two" and should be "two "
PASS Selection direction is correct.
PASS successfullyParsed is true

TEST COMPLETE
one two three
four five six
